Couple more renames
